User API and Authontication and authorization Done

+++ backend/controllers/goalController.js
@@ -1,11 +1,11 @@
 import asyncHandler from "express-async-handler";
-
 import Goal from "../models/goalModel.js";
+import User from "../models/userModel.js";
 
 //@desc get Goals
 //@Route GET /api/goals
 export const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({user: req.user.id});
   res.status(200).json(goals);
 });
 
@@ -18,6 +18,7 @@ export const setGoal = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id
   });
 
   res.status(200).json(goal);
@@ -33,6 +34,20 @@ export const updateGoal = asyncHandler(async (req, res) => {
     throw new Error("Goal not Found!");
   }
 
+  const user = await User.findById(req.user.id);
+
+  //check for user
+  if(!user){
+    res.status(401)
+    throw new Error("User not found!");
+  }
+
+  //make sure the logged in user matches the goal user
+  if(goal.user.toString() !== user.id){
+    res.status(401);
+    throw new Error("User not Authorized!");
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -42,11 +57,25 @@ export const updateGoal = asyncHandler(async (req, res) => {
 //@desc delete Goal
 //@Route POST /api/goals/:id
 export const deleteGoal = asyncHandler(async (req, res) => {
-  const goal = Goal.findByIdAndDelete(req.params.id);
+  const goal = await Goal.findByIdAndDelete(req.params.id);
 
   if (!goal) {
     res.status(400);
     throw new Error("Goal Not Found!");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  //check for user
+  if(!user){
+    res.status(401)
+    throw new Error("User not found!");
+  }
+
+  //make sure the logged in user matches the goal user
+  if(goal.user.toString() !== user.id){
+    res.status(401);
+    throw new Error("User not Authorized!");
   }
 
   await Goal.deleteOne();
